basic-autonomous-team/requirements.txt
pyautogen
ollama
markdown
